+++ tck/mp-jwt-embedded/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-org.apache.tomee.microprofile.jwt.MPJWTCDIExtension
+org.apache.tomee.microprofile.jwt.cdi.MPJWTCDIExtension
