--- tck/mp-jwt-embedded/src/main/resources/META-INF/services/org.apache.tomee.microprofile.jwt.JWTCallerPrincipalFactory
@@ -1,1 +1,0 @@
-org.apache.tomee.microprofile.jwt.DefaultJWTCallerPrincipalFactory
